# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow AI Project Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Projects ──────────────────────────────────────────────────────────────
...
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams
      responses:
        "200":
          description: Team list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /teams/{id}/members:
    post:
      operationId: addTeamMember
      tags: [teams]
      summary: Add a member to a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}/members/{userId}:
    delete:
      operationId: removeTeamMember
      tags: [teams]
      summary: Remove a member from a team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  # ── Comments ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, fullName, email, role, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [member, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        role:
          type: string
          enum: [member, admin]
        avatarUrl:
          type: string

    # ── Project schemas ───────────────────────────────────────────────────
...
    Team:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
    TeamInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    TeamMemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer

    # ── Comment schemas ───────────────────────────────────────────────────
...